// src/api/generated/models/dateWithSeenStatus.ts
/**
 * Generated by orval v7.18.0 🍺
 * Do not edit manually.
 * History Trivia Backend API
 * API for History Trivia Chronokwiz game
 * OpenAPI spec version: 1.0.0
 */

export interface DateWithSeenStatus {
  date?: string;
  isSeen?: boolean;
  /** User's score for this chronokwiz */
  score?: number;
}